e2e/repos/appA/helpers/pyproject.toml
[project]
name = "argo-generator"
version = "0.1.0"
description = "Tool to generate ArgoCD applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "jinja2>=3.1.5",
    "pytest>=8.3.4",
    "pyyaml>=6.0.2",
]
